Move nudge algorithms to NormalizedDurationRecord
