+++ src/components/datetime.rs
@@ -3,7 +3,7 @@
 use crate::{
     components::{
         calendar::Calendar,
-        duration::{round_relative_duration, TimeDuration},
+        duration::TimeDuration,
         Instant,
     },
     iso::{IsoDate, IsoDateSlots, IsoDateTime, IsoTime},
@@ -17,7 +17,7 @@ use tinystr::TinyAsciiStr;
 
 use super::{
     calendar::{CalendarDateLike, GetTemporalCalendar},
-    duration::{normalized::NormalizedTimeDuration, RelativeRoundResult},
+    duration::normalized::{NormalizedTimeDuration, RelativeRoundResult},
     Duration,
 };
 
@@ -154,8 +154,7 @@ impl DateTime {
 
         // 8. Return ? RoundRelativeDuration(diff, destEpochNs, dateTime, calendarRec, unset, largestUnit,
         // roundingIncrement, smallestUnit, roundingMode).
-        round_relative_duration(
-            &diff,
+        diff.round_relative_duration(
             dest_epoch_ns,
             self,
             None,
