+++ src/components/duration.rs
@@ -1,21 +1,15 @@
 //! This module implements `Duration` along with it's methods and components.
 
 use crate::{
-    components::DateTime,
-    iso::{IsoDate, IsoDateTime, IsoTime},
-    options::{
-        ArithmeticOverflow, RelativeTo, RoundingIncrement, TemporalRoundingMode, TemporalUnit,
-    },
-    rounding::{IncrementRounder, Round},
-    utils::NonZeroOrdering,
-    TemporalError, TemporalResult, TemporalUnwrap, NS_PER_DAY,
+    components::{DateTime, Time},
+    iso::{IsoDateTime, IsoTime},
+    options::{RelativeTo, RoundingIncrement, TemporalRoundingMode, TemporalUnit},
+    TemporalError, TemporalResult,
 };
 use ixdtf::parsers::{records::TimeDurationRecord, IsoDurationParser};
-use std::{num::NonZeroU64, str::FromStr};
+use std::str::FromStr;
 
 use self::normalized::{NormalizedDurationRecord, NormalizedTimeDuration};
-
-use super::{tz::TimeZone, Date, Time};
 
 mod date;
 pub(crate) mod normalized;
@@ -1030,600 +1024,4 @@ impl FromStr for Duration {
             )?,
         })
     }
-}
-
-// ==== New Nudge Algo Functions ====
-
-// TODO: Reorganize as needed.
-
-#[derive(Debug)]
-pub(crate) struct NudgeRecord {
-    normalized: NormalizedDurationRecord,
-    total: Option<i128>, // TODO: adjust
-    nudge_epoch_ns: i128,
-    expanded: bool,
-}
-
-// TODO: Add assertion into impl.
-// TODO: Add unit tests specifically for nudge_calendar_unit if possible.
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn nudge_calendar_unit(
-    sign: NonZeroOrdering,
-    duration: &NormalizedDurationRecord,
-    dest_epoch_ns: i128,
-    dt: &DateTime,
-    tz: Option<TimeZone>, // ???
-    increment: RoundingIncrement,
-    unit: TemporalUnit,
-    rounding_mode: TemporalRoundingMode,
-) -> TemporalResult<NudgeRecord> {
-    // NOTE: r2 may never be used...need to test.
-    let (r1, r2, start_duration, end_duration) = match unit {
-        // 1. If unit is "year", then
-        TemporalUnit::Year => {
-            // a. Let years be RoundNumberToIncrement(duration.[[Years]], increment, "trunc").
-            let years = IncrementRounder::from_potentially_negative_parts(
-                duration.date().years,
-                increment.as_extended_increment(),
-            )?
-            .round(TemporalRoundingMode::Trunc);
-            // b. Let r1 be years.
-            let r1 = years;
-            // c. Let r2 be years + increment × sign.
-            let r2 = years + i128::from(increment.get()) * i128::from(sign.get());
-            // d. Let startDuration be ? CreateNormalizedDurationRecord(r1, 0, 0, 0, ZeroTimeDuration()).
-            // e. Let endDuration be ? CreateNormalizedDurationRecord(r2, 0, 0, 0, ZeroTimeDuration()).
-            (
-                r1,
-                r2,
-                DateDuration::new(r1 as f64, 0.0, 0.0, 0.0)?,
-                DateDuration::new(r2 as f64, 0.0, 0.0, 0.0)?,
-            )
-        }
-        // 2. Else if unit is "month", then
-        TemporalUnit::Month => {
-            // a. Let months be RoundNumberToIncrement(duration.[[Months]], increment, "trunc").
-            let months = IncrementRounder::from_potentially_negative_parts(
-                duration.date().months,
-                increment.as_extended_increment(),
-            )?
-            .round(TemporalRoundingMode::Trunc);
-            // b. Let r1 be months.
-            let r1 = months;
-            // c. Let r2 be months + increment × sign.
-            let r2 = months + i128::from(increment.get()) * i128::from(sign.get());
-            // d. Let startDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], r1, 0, 0, ZeroTimeDuration()).
-            // e. Let endDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], r2, 0, 0, ZeroTimeDuration()).
-            (
-                r1,
-                r2,
-                DateDuration::new(duration.date().years, r1 as f64, 0.0, 0.0)?,
-                DateDuration::new(duration.date().years, r2 as f64, 0.0, 0.0)?,
-            )
-        }
-        // 3. Else if unit is "week", then
-        TemporalUnit::Week => {
-            // TODO: Reconcile potential overflow on years as i32. `ValidateDuration` requires years, months, weeks to be abs(x) <= 2^32
-            // a. Let isoResult1 be BalanceISODate(dateTime.[[Year]] + duration.[[Years]],
-            // dateTime.[[Month]] + duration.[[Months]], dateTime.[[Day]]).
-            let iso_one = IsoDate::balance(
-                dt.iso_year() + duration.date().years as i32,
-                i32::from(dt.iso_month()) + duration.date().months as i32,
-                i32::from(dt.iso_day()),
-            );
-
-            // b. Let isoResult2 be BalanceISODate(dateTime.[[Year]] + duration.[[Years]], dateTime.[[Month]] +
-            // duration.[[Months]], dateTime.[[Day]] + duration.[[Days]]).
-            let iso_two = IsoDate::balance(
-                dt.iso_year() + duration.date().years as i32,
-                i32::from(dt.iso_month()) + duration.date().months as i32,
-                i32::from(dt.iso_day()) + duration.date().days as i32,
-            );
-
-            // c. Let weeksStart be ! CreateTemporalDate(isoResult1.[[Year]], isoResult1.[[Month]], isoResult1.[[Day]],
-            // calendarRec.[[Receiver]]).
-            let weeks_start = Date::new(
-                iso_one.year,
-                iso_one.month.into(),
-                iso_one.day.into(),
-                dt.calendar().clone(),
-                ArithmeticOverflow::Reject,
-            )?;
-
-            // d. Let weeksEnd be ! CreateTemporalDate(isoResult2.[[Year]], isoResult2.[[Month]], isoResult2.[[Day]],
-            // calendarRec.[[Receiver]]).
-            let weeks_end = Date::new(
-                iso_two.year,
-                iso_two.month.into(),
-                iso_two.day.into(),
-                dt.calendar().clone(),
-                ArithmeticOverflow::Reject,
-            )?;
-
-            // e. Let untilOptions be OrdinaryObjectCreate(null).
-            // f. Perform ! CreateDataPropertyOrThrow(untilOptions, "largestUnit", "week").
-            // g. Let untilResult be ? DifferenceDate(calendarRec, weeksStart, weeksEnd, untilOptions).
-            let until_result =
-                weeks_start.internal_diff_date(&weeks_end, TemporalUnit::Week)?;
-
-            // h. Let weeks be RoundNumberToIncrement(duration.[[Weeks]] + untilResult.[[Weeks]], increment, "trunc").
-            let weeks = IncrementRounder::from_potentially_negative_parts(
-                duration.date().weeks + until_result.weeks(),
-                increment.as_extended_increment(),
-            )?
-            .round(TemporalRoundingMode::Trunc);
-
-            // i. Let r1 be weeks.
-            let r1 = weeks;
-            // j. Let r2 be weeks + increment × sign.
-            let r2 = weeks + i128::from(increment.get()) * i128::from(sign.get());
-            // k. Let startDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], r1, 0, ZeroTimeDuration()).
-            // l. Let endDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], r2, 0, ZeroTimeDuration()).
-            (
-                r1,
-                r2,
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months,
-                    r1 as f64,
-                    0.0,
-                )?,
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months,
-                    r2 as f64,
-                    0.0,
-                )?,
-            )
-        }
-        TemporalUnit::Day => {
-            // 4. Else,
-            // a. Assert: unit is "day".
-            // b. Let days be RoundNumberToIncrement(duration.[[Days]], increment, "trunc").
-            let days = IncrementRounder::from_potentially_negative_parts(
-                duration.date().days,
-                increment.as_extended_increment(),
-            )?
-            .round(TemporalRoundingMode::Trunc);
-            // c. Let r1 be days.
-            let r1 = days;
-            // d. Let r2 be days + increment × sign.
-            let r2 = days + i128::from(increment.get()) * i128::from(sign.get());
-            // e. Let startDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], duration.[[Weeks]], r1, ZeroTimeDuration()).
-            // f. Let endDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], duration.[[Weeks]], r2, ZeroTimeDuration()).
-            (
-                r1,
-                r2,
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months,
-                    duration.date().weeks,
-                    r1 as f64,
-                )?,
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months,
-                    duration.date().weeks,
-                    r2 as f64,
-                )?,
-            )
-        }
-        _ => unreachable!(), // TODO: potentially reject with range error?
-    };
-
-    // 5. Let start be ? AddDateTime(dateTime.[[Year]], dateTime.[[Month]], dateTime.[[Day]], dateTime.[[Hour]], dateTime.[[Minute]],
-    // dateTime.[[Second]], dateTime.[[Millisecond]], dateTime.[[Microsecond]], dateTime.[[Nanosecond]], calendarRec,
-    // startDuration.[[Years]], startDuration.[[Months]], startDuration.[[Weeks]], startDuration.[[Days]], startDuration.[[NormalizedTime]], undefined).
-    let start = dt.iso.add_date_duration(
-        dt.calendar().clone(),
-        &start_duration,
-        NormalizedTimeDuration::default(),
-        None,
-    )?;
-
-    // 6. Let end be ? AddDateTime(dateTime.[[Year]], dateTime.[[Month]], dateTime.[[Day]], dateTime.[[Hour]],
-    // dateTime.[[Minute]], dateTime.[[Second]], dateTime.[[Millisecond]], dateTime.[[Microsecond]],
-    // dateTime.[[Nanosecond]], calendarRec, endDuration.[[Years]], endDuration.[[Months]], endDuration.[[Weeks]],
-    // endDuration.[[Days]], endDuration.[[NormalizedTime]], undefined).
-    let end = dt.iso.add_date_duration(
-        dt.calendar().clone(),
-        &end_duration,
-        NormalizedTimeDuration::default(),
-        None,
-    )?;
-
-    // 7. If timeZoneRec is unset, then
-    let (start_epoch_ns, end_epoch_ns) = if tz.is_none() {
-        // TODO: Test valid range of EpochNanoseconds in order to add `expect` over `unwrap_or`
-        // a. Let startEpochNs be GetUTCEpochNanoseconds(start.[[Year]], start.[[Month]], start.[[Day]], start.[[Hour]], start.[[Minute]], start.[[Second]], start.[[Millisecond]], start.[[Microsecond]], start.[[Nanosecond]]).
-        // b. Let endEpochNs be GetUTCEpochNanoseconds(end.[[Year]], end.[[Month]], end.[[Day]], end.[[Hour]], end.[[Minute]], end.[[Second]], end.[[Millisecond]], end.[[Microsecond]], end.[[Nanosecond]]).
-        (
-            start.as_nanoseconds(0.0).unwrap_or(0),
-            end.as_nanoseconds(0.0).unwrap_or(0),
-        )
-    // 8. Else,
-    } else {
-        // a. Let startDateTime be ! CreateTemporalDateTime(start.[[Year]], start.[[Month]], start.[[Day]],
-        // start.[[Hour]], start.[[Minute]], start.[[Second]], start.[[Millisecond]], start.[[Microsecond]],
-        // start.[[Nanosecond]], calendarRec.[[Receiver]]).
-        // b. Let startInstant be ? GetInstantFor(timeZoneRec, startDateTime, "compatible").
-        // c. Let startEpochNs be startInstant.[[Nanoseconds]].
-        // d. Let endDateTime be ! CreateTemporalDateTime(end.[[Year]], end.[[Month]], end.[[Day]], end.[[Hour]], end.[[Minute]], end.[[Second]], end.[[Millisecond]], end.[[Microsecond]], end.[[Nanosecond]], calendarRec.[[Receiver]]).
-        // e. Let endInstant be ? GetInstantFor(timeZoneRec, endDateTime, "compatible").
-        // f. Let endEpochNs be endInstant.[[Nanoseconds]].
-        return Err(TemporalError::general(
-            "TimeZone handling not yet implemented.",
-        ));
-    };
-
-    // 9. If endEpochNs = startEpochNs, throw a RangeError exception.
-    if end_epoch_ns == start_epoch_ns {
-        return Err(
-            TemporalError::range().with_message("endEpochNs cannot be equal to startEpochNs")
-        );
-    }
-
-    // TODO: Add early RangeError steps
-
-    // NOTE: Below is removed in place of using `IncrementRounder`
-    // 10. If sign < 0, let isNegative be negative; else let isNegative be positive.
-    // 11. Let unsignedRoundingMode be GetUnsignedRoundingMode(roundingMode, isNegative).
-
-    // TODO: Step 12..13 could be problematic...need tests
-    // and verify, or completely change the approach involved.
-    // 12. Let progress be (destEpochNs - startEpochNs) / (endEpochNs - startEpochNs).
-    // 13. Let total be r1 + progress × increment × sign.
-
-    // TODO: Remove if invalid
-    // NOTE: Changes to 12 -> 12. Let progress be ((destEpochNs - startEpochNs) * increment) / (endEpochNs - startEpochNs).
-    // let progress = ((dest_epoch_ns - start_epoch_ns) * i128::from(increment.get()))
-    //     .div_euclid(end_epoch_ns - start_epoch_ns);
-    // NOTE: Changes to 13 -> 13. let total be r1 + progress * sign.
-    // let total = r1 + progress * i128::from(sign);
-
-    let progress = (dest_epoch_ns - start_epoch_ns) as f64 / (end_epoch_ns - start_epoch_ns) as f64;
-    let total = r1 as f64 + progress * increment.get() as f64 * f64::from(sign.get());
-
-    // TODO: Test and verify that `IncrementRounder` handles the below case.
-    // NOTE(nekevss): Below will not return the calculated r1 or r2, so it is imporant to not use
-    // the result beyond determining rounding direction.
-    // 14. NOTE: The above two steps cannot be implemented directly using floating-point arithmetic.
-    // This division can be implemented as if constructing Normalized Time Duration Records for the denominator
-    // and numerator of total and performing one division operation with a floating-point result.
-    // 15. Let roundedUnit be ApplyUnsignedRoundingMode(total, r1, r2, unsignedRoundingMode).
-    let rounded_unit = IncrementRounder::from_potentially_negative_parts(
-        total,
-        increment.as_extended_increment(),
-    )?
-    .round(rounding_mode);
-
-    // 16. If roundedUnit - total < 0, let roundedSign be -1; else let roundedSign be 1.
-    // 19. Return Duration Nudge Result Record { [[Duration]]: resultDuration, [[Total]]: total, [[NudgedEpochNs]]: nudgedEpochNs, [[DidExpandCalendarUnit]]: didExpandCalendarUnit }.
-    // 17. If roundedSign = sign, then
-    if rounded_unit == r2 {
-        // a. Let didExpandCalendarUnit be true.
-        // b. Let resultDuration be endDuration.
-        // c. Let nudgedEpochNs be endEpochNs.
-        Ok(NudgeRecord {
-            normalized: NormalizedDurationRecord::new(
-                end_duration,
-                NormalizedTimeDuration::default(),
-            )?,
-            total: Some(total as i128),
-            nudge_epoch_ns: end_epoch_ns,
-            expanded: true,
-        })
-    // 18. Else,
-    } else {
-        // a. Let didExpandCalendarUnit be false.
-        // b. Let resultDuration be startDuration.
-        // c. Let nudgedEpochNs be startEpochNs.
-        Ok(NudgeRecord {
-            normalized: NormalizedDurationRecord::new(
-                start_duration,
-                NormalizedTimeDuration::default(),
-            )?,
-            total: Some(total as i128),
-            nudge_epoch_ns: start_epoch_ns,
-            expanded: false,
-        })
-    }
-}
-
-#[inline]
-fn nudge_to_zoned_time() -> TemporalResult<NudgeRecord> {
-    // TODO: Implement
-    Err(TemporalError::general("Not yet implemented."))
-}
-
-#[inline]
-fn nudge_to_day_or_time(
-    duration: &NormalizedDurationRecord,
-    dest_epoch_ns: i128,
-    largest_unit: TemporalUnit,
-    increment: RoundingIncrement,
-    smallest_unit: TemporalUnit,
-    rounding_mode: TemporalRoundingMode,
-) -> TemporalResult<NudgeRecord> {
-    // 1. Assert: The value in the "Category" column of the row of Table 22 whose "Singular" column contains smallestUnit, is time.
-    // 2. Let norm be ! Add24HourDaysToNormalizedTimeDuration(duration.[[NormalizedTime]], duration.[[Days]]).
-    let norm = duration.norm().add_days(duration.date().days as i64)?;
-
-    // 3. Let unitLength be the value in the "Length in Nanoseconds" column of the row of Table 22 whose "Singular" column contains smallestUnit.
-    let unit_length = smallest_unit.as_nanoseconds().temporal_unwrap()?;
-    // 4. Let total be DivideNormalizedTimeDuration(norm, unitLength).
-    let total = norm.divide(unit_length as i64);
-
-    // 5. Let roundedNorm be ? RoundNormalizedTimeDurationToIncrement(norm, unitLength × increment, roundingMode).
-    let rounded_norm = norm.round(
-        unsafe {
-            NonZeroU64::new_unchecked(unit_length)
-                .checked_mul(increment.as_extended_increment())
-                .temporal_unwrap()?
-        },
-        rounding_mode,
-    )?;
-
-    // 6. Let diffNorm be ! SubtractNormalizedTimeDuration(roundedNorm, norm).
-    let diff_norm = rounded_norm.checked_sub(&norm)?;
-
-    // 7. Let wholeDays be truncate(DivideNormalizedTimeDuration(norm, nsPerDay)).
-    let whole_days = norm.divide(NS_PER_DAY as i64);
-
-    // 8. Let roundedFractionalDays be DivideNormalizedTimeDuration(roundedNorm, nsPerDay).
-    let (rounded_whole_days, rounded_remainder) = rounded_norm.div_rem(NS_PER_DAY);
-
-    // 9. Let roundedWholeDays be truncate(roundedFractionalDays).
-    // 10. Let dayDelta be roundedWholeDays - wholeDays.
-    let delta = rounded_whole_days - whole_days;
-    // 11. If dayDelta < 0, let dayDeltaSign be -1; else if dayDelta > 0, let dayDeltaSign be 1; else let dayDeltaSign be 0.
-    // 12. If dayDeltaSign = NormalizedTimeDurationSign(norm), let didExpandDays be true; else let didExpandDays be false.
-    let did_expand_days = delta.signum() == norm.sign().into();
-
-    // 13. Let nudgedEpochNs be AddNormalizedTimeDurationToEpochNanoseconds(diffNorm, destEpochNs).
-    let nudged_ns = diff_norm.0 + dest_epoch_ns;
-
-    // 14. Let days be 0.
-    let mut days = 0;
-    // 15. Let remainder be roundedNorm.
-    let mut remainder = rounded_norm;
-    // 16. If LargerOfTwoTemporalUnits(largestUnit, "day") is largestUnit, then
-    if largest_unit.max(TemporalUnit::Day) == largest_unit {
-        // a. Set days to roundedWholeDays.
-        days = rounded_whole_days;
-        // b. Set remainder to remainder(roundedFractionalDays, 1) × nsPerDay.
-        remainder = NormalizedTimeDuration(rounded_remainder);
-    }
-    // 17. Let resultDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], duration.[[Weeks]], days, remainder).
-    let result_duration = NormalizedDurationRecord::new(
-        DateDuration::new(
-            duration.date().years,
-            duration.date().months,
-            duration.date().weeks,
-            days as f64,
-        )?,
-        remainder,
-    )?;
-    // 18. Return Duration Nudge Result Record { [[Duration]]: resultDuration, [[Total]]: total,
-    // [[NudgedEpochNs]]: nudgedEpochNs, [[DidExpandCalendarUnit]]: didExpandDays }.
-    Ok(NudgeRecord {
-        normalized: result_duration,
-        total: Some(total),
-        nudge_epoch_ns: nudged_ns,
-        expanded: did_expand_days,
-    })
-}
-
-// 7.5.43 BubbleRelativeDuration ( sign, duration, nudgedEpochNs, dateTime, calendarRec, timeZoneRec, largestUnit, smallestUnit )
-#[inline]
-#[allow(clippy::too_many_arguments)]
-fn bubble_relative_duration(
-    sign: NonZeroOrdering,
-    duration: &NormalizedDurationRecord,
-    nudge_epoch_ns: i128,
-    date_time: &DateTime,
-    tz: Option<TimeZone>,
-    largest_unit: TemporalUnit,
-    smallest_unit: TemporalUnit,
-) -> TemporalResult<NormalizedDurationRecord> {
-    // Assert: The value in the "Category" column of the row of Table 22 whose "Singular" column contains largestUnit, is date.
-    // 2. Assert: The value in the "Category" column of the row of Table 22 whose "Singular" column contains smallestUnit, is date.
-    let mut duration = *duration;
-    // 3. If smallestUnit is "year", return duration.
-    if smallest_unit == TemporalUnit::Year {
-        return Ok(duration);
-    }
-
-    // NOTE: Invert ops as Temporal Proposal table is inverted (i.e. Year = 0 ... Nanosecond = 9)
-    // 4. Let largestUnitIndex be the ordinal index of the row of Table 22 whose "Singular" column contains largestUnit.
-    // 5. Let smallestUnitIndex be the ordinal index of the row of Table 22 whose "Singular" column contains smallestUnit.
-    // 6. Let unitIndex be smallestUnitIndex - 1.
-    let mut unit = smallest_unit + 1;
-    // 7. Let done be false.
-    // 8. Repeat, while unitIndex ≤ largestUnitIndex and done is false,
-    while unit != TemporalUnit::Auto && unit <= largest_unit {
-        // a. Let unit be the value in the "Singular" column of Table 22 in the row whose ordinal index is unitIndex.
-        // b. If unit is not "week", or largestUnit is "week", then
-        if unit == TemporalUnit::Week || largest_unit != TemporalUnit::Week {
-            unit = unit + 1;
-            continue;
-        }
-
-        let end_duration = match unit {
-            // i. If unit is "year", then
-            TemporalUnit::Year => {
-                // 1. Let years be duration.[[Years]] + sign.
-                // 2. Let endDuration be ? CreateNormalizedDurationRecord(years, 0, 0, 0, ZeroTimeDuration()).
-                DateDuration::new(duration.date().years + f64::from(sign.get()), 0.0, 0.0, 0.0)?
-            }
-            // ii. Else if unit is "month", then
-            TemporalUnit::Month => {
-                // 1. Let months be duration.[[Months]] + sign.
-                // 2. Let endDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], months, 0, 0, ZeroTimeDuration()).
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months + f64::from(sign.get()),
-                    0.0,
-                    0.0,
-                )?
-            }
-            // iii. Else if unit is "week", then
-            TemporalUnit::Week => {
-                // 1. Let weeks be duration.[[Weeks]] + sign.
-                // 2. Let endDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], weeks, 0, ZeroTimeDuration()).
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months,
-                    duration.date().weeks + f64::from(sign.get()),
-                    0.0,
-                )?
-            }
-            // iv. Else,
-            TemporalUnit::Day => {
-                // 1. Assert: unit is "day".
-                // 2. Let days be duration.[[Days]] + sign.
-                // 3. Let endDuration be ? CreateNormalizedDurationRecord(duration.[[Years]], duration.[[Months]], duration.[[Weeks]], days, ZeroTimeDuration()).
-                DateDuration::new(
-                    duration.date().years,
-                    duration.date().months,
-                    duration.date().weeks,
-                    duration.date().days + f64::from(sign.get()),
-                )?
-            }
-            _ => unreachable!(),
-        };
-
-        // v. Let end be ? AddDateTime(dateTime.[[Year]], dateTime.[[Month]], dateTime.[[Day]], dateTime.[[Hour]], dateTime.[[Minute]],
-        // dateTime.[[Second]], dateTime.[[Millisecond]], dateTime.[[Microsecond]], dateTime.[[Nanosecond]], calendarRec,
-        // endDuration.[[Years]], endDuration.[[Months]], endDuration.[[Weeks]], endDuration.[[Days]], endDuration.[[NormalizedTime]], undefined).
-        let end = date_time.iso.add_date_duration(
-            date_time.calendar().clone(),
-            &end_duration,
-            NormalizedTimeDuration::default(),
-            None,
-        )?;
-
-        // vi. If timeZoneRec is unset, then
-        let end_epoch_ns = if let Some(ref _tz) = tz {
-            // 1. Let endDateTime be ! CreateTemporalDateTime(end.[[Year]], end.[[Month]], end.[[Day]],
-            // end.[[Hour]], end.[[Minute]], end.[[Second]], end.[[Millisecond]], end.[[Microsecond]],
-            // end.[[Nanosecond]], calendarRec.[[Receiver]]).
-            // 2. Let endInstant be ? GetInstantFor(timeZoneRec, endDateTime, "compatible").
-            // 3. Let endEpochNs be endInstant.[[Nanoseconds]].
-            return Err(TemporalError::general("Not yet implemented."));
-        // vii. Else,
-        } else {
-            // 1. Let endEpochNs be GetUTCEpochNanoseconds(end.[[Year]], end.[[Month]], end.[[Day]], end.[[Hour]],
-            // end.[[Minute]], end.[[Second]], end.[[Millisecond]], end.[[Microsecond]], end.[[Nanosecond]]).
-            end.as_nanoseconds(0.0).temporal_unwrap()?
-        };
-        // viii. Let beyondEnd be nudgedEpochNs - endEpochNs.
-        let beyond_end = nudge_epoch_ns - end_epoch_ns;
-        // ix. If beyondEnd < 0, let beyondEndSign be -1; else if beyondEnd > 0, let beyondEndSign be 1; else let beyondEndSign be 0.
-        // x. If beyondEndSign ≠ -sign, then
-        if beyond_end.signum() != -i128::from(sign.get()) {
-            // 1. Set duration to endDuration.
-            duration = NormalizedDurationRecord::from_date_duration(end_duration)?;
-        // xi. Else,
-        } else {
-            // 1. Set done to true.
-            break;
-        }
-        // c. Set unitIndex to unitIndex - 1.
-        unit = unit + 1;
-    }
-
-    Ok(duration)
-}
-
-pub(crate) type RelativeRoundResult = (Duration, Option<i128>);
-
-// 7.5.44 RoundRelativeDuration ( duration, destEpochNs, dateTime, calendarRec, timeZoneRec, largestUnit, increment, smallestUnit, roundingMode )
-#[allow(clippy::too_many_arguments)]
-#[inline]
-pub(crate) fn round_relative_duration(
-    duration: &NormalizedDurationRecord,
-    dest_epoch_ns: i128,
-    dt: &DateTime,
-    tz: Option<TimeZone>,
-    largest_unit: TemporalUnit,
-    increment: RoundingIncrement,
-    smallest_unit: TemporalUnit,
-    rounding_mode: TemporalRoundingMode,
-) -> TemporalResult<RelativeRoundResult> {
-    // 1. Let irregularLengthUnit be false.
-    // 2. If IsCalendarUnit(smallestUnit) is true, set irregularLengthUnit to true.
-    // 3. If timeZoneRec is not unset and smallestUnit is "day", set irregularLengthUnit to true.
-    let irregular_unit =
-        smallest_unit.is_calendar_unit() || (tz.is_some() && smallest_unit == TemporalUnit::Day);
-
-    // 4. If DurationSign(duration.[[Years]], duration.[[Months]], duration.[[Weeks]], duration.[[Days]], NormalizedTimeDurationSign(duration.[[NormalizedTime]]), 0, 0, 0, 0, 0) < 0, let sign be -1; else let sign be 1.
-    let sign = NonZeroOrdering::new(duration.sign()?.cmp(&0));
-
-    // 5. If irregularLengthUnit is true, then
-    let nudge_result = if irregular_unit {
-        // a. Let nudgeResult be ? NudgeToCalendarUnit(sign, duration, destEpochNs, dateTime, calendarRec, timeZoneRec, increment, smallestUnit, roundingMode).
-        nudge_calendar_unit(
-            sign,
-            duration,
-            dest_epoch_ns,
-            dt,
-            tz.clone(),
-            increment,
-            smallest_unit,
-            rounding_mode,
-        )?
-    // 6. Else if timeZoneRec is not unset, then
-    } else if let Some(ref _tz) = tz {
-        // a. Let nudgeResult be ? NudgeToZonedTime(sign, duration, dateTime, calendarRec, timeZoneRec, increment, smallestUnit, roundingMode).
-        nudge_to_zoned_time()?
-    // 7. Else,
-    } else {
-        // a. Let nudgeResult be ? NudgeToDayOrTime(duration, destEpochNs, largestUnit, increment, smallestUnit, roundingMode).
-        nudge_to_day_or_time(
-            duration,
-            dest_epoch_ns,
-            largest_unit,
-            increment,
-            smallest_unit,
-            rounding_mode,
-        )?
-    };
-
-    // 8. Set duration to nudgeResult.[[Duration]].
-    let mut duration = nudge_result.normalized;
-
-    // 9. If nudgeResult.[[DidExpandCalendarUnit]] is true and smallestUnit is not "week", then
-    if nudge_result.expanded && smallest_unit != TemporalUnit::Week {
-        // a. Let startUnit be LargerOfTwoTemporalUnits(smallestUnit, "day").
-        let start_unit = smallest_unit.max(TemporalUnit::Day);
-        // b. Set duration to ? BubbleRelativeDuration(sign, duration, nudgeResult.[[NudgedEpochNs]], dateTime, calendarRec, timeZoneRec, largestUnit, startUnit).
-        duration = bubble_relative_duration(
-            sign,
-            &duration,
-            nudge_result.nudge_epoch_ns,
-            dt,
-            tz,
-            largest_unit,
-            start_unit,
-        )?
-    };
-
-    // 10. If IsCalendarUnit(largestUnit) is true or largestUnit is "day", then
-    let largest_unit = if largest_unit.is_calendar_unit() || largest_unit == TemporalUnit::Day {
-        // a. Set largestUnit to "hour".
-        TemporalUnit::Hour
-    } else {
-        largest_unit
-    };
-
-    // 11. Let balanceResult be ? BalanceTimeDuration(duration.[[NormalizedTime]], largestUnit).
-    let balance_result = TimeDuration::from_normalized(duration.0 .1, largest_unit)?;
-
-    // TODO: Need to validate the below.
-    // 12. Return the Record { [[Duration]]: CreateDurationRecord(duration.[[Years]], duration.[[Months]], duration.[[Weeks]], duration.[[Days]], balanceResult.[[Hours]], balanceResult.[[Minutes]], balanceResult.[[Seconds]], balanceResult.[[Milliseconds]], balanceResult.[[Microseconds]], balanceResult.[[Nanoseconds]]), [[Total]]: nudgeResult.[[Total]]  }.
-    Ok((
-        Duration::new_unchecked(duration.date(), balance_result.1),
-        nudge_result.total,
-    ))
 }
